+++ CHANGELOG.rst
@@ -9,6 +9,12 @@ Changelog
 latest
 ------
 
+
+- WIP: Memory improvements for parallel execution in simulations.
+
+  - ``get_model`` is done on the fly, not stored in a dict any longer
+    (``simulation._dict_model`` does not exist any longer).
+
 - Bug fixes:
 
   - Track order when saving to hdf5.